+++ tests/building.R
@@ -1,4 +1,4 @@
-# 
+#
 # library(rstudioapi)
 # script_path <- getActiveDocumentContext()$path
 # print(script_path)