+++ NEWS.md
@@ -1,3 +1,11 @@
+# EFAfactors 1.1.0
+
+-   Added   -  Added factor analysis eigenvalues for `EFAscreet`, `Hull`, and `KGC`.
+-   Fixed   -  Corrected the erroneous descriptions in the manual.
+-   Fixed   -  Corrected the error in the Hull plot.
+-   Fixed   -  Corrected the calculation method for reference eigenvalues in factor analysis based PA.
+
 # EFAfactors 1.0.0
 
--   Initial release
+-   Initial release.
+